pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "rasyn"
version = "0.1.0"
description = "Hybrid retrosynthesis AI: graph edit head + LLM generator + process-aware ranking"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torch-geometric>=2.4.0",
    "rdkit>=2022.9.5",
    "rdchiral>=1.0.0",
    "transformers>=4.42.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "deepspeed>=0.12.0",
    "numpy>=1.23.0",
    "pandas>=1.5.0",
    "pyarrow>=12.0.0",
    "pyyaml>=6.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
include = ["rasyn*"]
